Refactor validation context and profile rule handling
- Updated `Writer` struct to use `u64` for segment count to prevent overflow.
- Modified `Validator` trait to include `ValidationRuleContext` in `validate_batch` method.
- Introduced a common demo orders profile pack for validation tests.
- Updated tests to utilize the new profile pack and ensure consistent validation behavior.
- Enhanced diagnostics snapshots to use the common profile pack.
- Improved validation tests to check for message type handling and error reporting.
- Updated `deny.toml` to enforce stricter dependency policies, including unlicensed and copyleft crates.
- Adjusted script comments for clarity on benchmarking steps.

+++ crates/edifact-rs/src/writer.rs
@@ -10,7 +10,9 @@ use std::io::Write;
 pub struct Writer<W: Write> {
     inner: W,
     ssa: ServiceStringAdvice,
-    segment_count: u32,
+    /// Running count of segments written.  `u64` to prevent silent overflow on
+    /// pathological inputs (a `u32` would wrap after ~4 billion segments).
+    segment_count: u64,
 }
 
 impl<W: Write> Writer<W> {
@@ -156,7 +158,8 @@ impl<W: Write> Writer<W> {
     }
 
     /// Number of segments written so far.
-    pub fn segment_count(&self) -> u32 {
+    /// Returns the total number of segments written so far.
+    pub fn segment_count(&self) -> u64 {
         self.segment_count
     }
